streamlit_dashboard/requirements.txt
streamlit
pandas
altair
matplotlib
numpy
scipy
google-genai
python-dotenv